Moved charts into separate directory
